chore: separate shared ui in new form component into another component
